backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pymongo==4.6.2
motor==3.3.2
httpx==0.27.0
beautifulsoup4==4.12.3
google-generativeai==0.5.4
python-dotenv==1.0.1
pydantic[email]==2.6.4
duckduckgo-search==5.3.0
lxml==5.1.0
requests==2.31.0
aiohttp==3.9.3
tenacity==8.2.3
langdetect==1.0.9
